Cleaned code, moved distance Calculation method to Utils class. Also fixed minor issue where when something is first searched normally and then with a filter, results are added to the bottom of the previous search by clearing the list before searching with a filter
